Atualizações de Segurança como excluir usuário e validações de login/registro

+++ src/main/java/com/LucasVicentee/ProjetoGestaoFinanceira_BackEnd/service/user/UserService.java
@@ -5,6 +5,8 @@ import com.LucasVicentee.ProjetoGestaoFinanceira_BackEnd.dto.user.UserResponseDT
 import com.LucasVicentee.ProjetoGestaoFinanceira_BackEnd.dto.user.UserUpdateDTO;
 import com.LucasVicentee.ProjetoGestaoFinanceira_BackEnd.exceptions.RequiredObjectIsNullException;
 import com.LucasVicentee.ProjetoGestaoFinanceira_BackEnd.exceptions.ResourceNotFoundException;
+import com.LucasVicentee.ProjetoGestaoFinanceira_BackEnd.exceptions.user.EmailAlreadyInUseException;
+import com.LucasVicentee.ProjetoGestaoFinanceira_BackEnd.exceptions.user.ForbiddenException;
 import com.LucasVicentee.ProjetoGestaoFinanceira_BackEnd.mapper.user.UserMapper;
 import com.LucasVicentee.ProjetoGestaoFinanceira_BackEnd.model.user.User;
 import com.LucasVicentee.ProjetoGestaoFinanceira_BackEnd.repository.user.UserRepository;
@@ -33,6 +35,10 @@ public class UserService {
         if (user == null) throw new RequiredObjectIsNullException();
         logger.info("Creating a new User!");
 
+        if (repository.existsByEmail(user.getEmail())) {
+            throw new EmailAlreadyInUseException("O E-mail " + user.getEmail() + "já está cadastrado.");
+        }
+
         User entity = mapper.toEntity(user);
         entity.setPassword(passwordEncoder.encode(entity.getPassword())); // Pega a senha passada "crua" pelo usuário e codifica
         User savedEntity = repository.save(entity); // Salva no repositório
@@ -43,15 +49,19 @@ public class UserService {
         return passwordEncoder.matches(rawPassWord, storedHashedPassword); // Verifica se a senha que o usuário digitar é igual ao do armazenado no banco em Hash no momento do Login
     }
 
-    public UserResponseDTO updatePartially(Long id, UserUpdateDTO user) {
+    public UserResponseDTO updatePartially(Long targetId, Long loggedInUserId, UserUpdateDTO user) {
 
         if (user == null) {
             throw new RequiredObjectIsNullException("The request body cannot be null");
         }
 
+        if (!targetId.equals(loggedInUserId)) {
+            throw new ForbiddenException("Você não tem permissão para modificar o perfil de outro usuário.");
+        }
+
         logger.info("Updating partially user with ID: " + user.getId());
 
-        User entity = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Records not found here!"));
+        User entity = repository.findById(targetId).orElseThrow(() -> new ResourceNotFoundException("Records not found here!"));
 
         if (user.getFirstName() != null) {
             entity.setFirstName(user.getFirstName());
@@ -74,10 +84,15 @@ public class UserService {
         return mapper.toDTO(updatedEntity);
     }
 
-    public void delete(Long id) {
+    public void delete(Long targetId, Long loggedInUserId) {
         logger.info("Deleting a User!");
 
-        User entity = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("ID Not Found!"));
+        if (!targetId.equals(loggedInUserId)) {
+            throw new ForbiddenException("Você não tem permissão para deletar o perfil de outro usuário.");
+        }
+
+        User entity = repository.findById(targetId).orElseThrow(() -> new ResourceNotFoundException("ID do Usuário não encontrado para exclusão!"));
+
         repository.delete(entity);
     }
 }
